# set workers with env var 