Allow setting workers for `parallelize` calls with an env var (#6).  It speeds up forward and backwards matrix multiplication a lot.  It didn't do much for conv2d though.  Mojo allows you to set env vars with `mojo -D WORKERS=4 ...` so it was cleaner than adding a CLI arg to each network example.

It doesn't change behaviour if the env var is not set so the change is purely additive in terms of the performance.  There may be systems where always parallelising by the number of iterations (the current setup) is optimal.  Tests below on a 4 core machine show significant speed up when specifying 4 workers.  

Using `transformer.mojo` for benchmarks since matrix multiplication is heavily used in that network.  With this change, forward `mul` goes from 21.58 -> 16.81 and backward `mul` goes from 126.72 -> 78.83.  Full timings below.

old `mojo transformer.mojo`

```
Average time spent in each operation per forward pass:
mul:            21.574972152709961 ms  ( 95.630821228027344 % )
add:            0.11080162227153778 ms  ( 0.49112695455551147 % )
relu:           0.018484519794583321 ms  ( 0.081932425498962402 % )
softmax:        0.64765501022338867 ms  ( 2.8707239627838135 % )
ce:             0.030097030103206635 ms  ( 0.13340474665164948 % )
reshape:        0.085219144821166992 ms  ( 0.37773290276527405 % )
transpose:      0.073509320616722107 ms  ( 0.32582923769950867 % )
copy:           0.01995040662586689 ms  ( 0.088429957628250122 % )

Average time spent in each operation per backward pass:
mul:            126.72743225097656 ms  ( 98.134323120117188 % )
add:            0.31367659568786621 ms  ( 0.24290274083614349 % )
relu:           0.081546954810619354 ms  ( 0.063147768378257751 % )
softmax:        0.79221165180206299 ms  ( 0.61346739530563354 % )
ce:             0.016986474394798279 ms  ( 0.013153868727385998 % )
reshape:        0.49982357025146484 ms  ( 0.38704994320869446 % )
transpose:      0.65263009071350098 ms  ( 0.50537919998168945 % )
copy:           0.052403680980205536 ms  ( 0.040580000728368759 % )
```

this PR `mojo -D WORKERS=4 transformer.mojo`
```
Average time spent in each operation per forward pass:
mul:            16.811895370483398 ms  ( 94.395881652832031 % )
add:            0.10715033859014511 ms  ( 0.60163050889968872 % )
relu:           0.028773274272680283 ms  ( 0.16155692934989929 % )
softmax:        0.64294660091400146 ms  ( 3.6100337505340576 % )
ce:             0.025745084509253502 ms  ( 0.14455416798591614 % )
reshape:        0.10363511741161346 ms  ( 0.5818932056427002 % )
transpose:      0.071380391716957092 ms  ( 0.40078851580619812 % )
copy:           0.018463574349880219 ms  ( 0.10366976261138916 % )

Average time spent in each operation per backward pass:
mul:            78.828399658203125 ms  ( 97.068397521972656 % )
add:            0.30091404914855957 ms  ( 0.37054216861724854 % )
relu:           0.1118263304233551 ms  ( 0.13770169019699097 % )
softmax:        0.78190362453460693 ms  ( 0.96282732486724854 % )
ce:             0.016558287665247917 ms  ( 0.02038968913257122 % )
reshape:        0.48039546608924866 ms  ( 0.59155356884002686 % )
transpose:      0.64519572257995605 ms  ( 0.79448670148849487 % )
copy:           0.04393550381064415 ms  ( 0.05410168319940567 % )
```